Reorganize sentry loader
