+++ src/core/DiscordClient.ts
@@ -1,12 +1,17 @@
-import { ActivityType, Client, GatewayIntentBits, Partials } from "discord.js";
+import { ActivityType, Client, GatewayIntentBits, Partials, DiscordAPIError } from "discord.js";
+import { APIErrors } from "../utils/discordErrorCode";
 import config from '../config.json';
 import { PrismaClient } from "@prisma/client";
 import { RedisClientType, createClient } from "redis";
 import { connect, Connection } from "amqplib";
 import { eventLogger } from "../utils/eventLogger";
-import { DiscordEvents } from "../events/discordEvents";
-import { RedisEvents } from "../events/redisEvents";
-import { AMQPEvents } from "../events/amqpEvents";
+import { DiscordEvents, RedisEvents, AMQPEvents } from "../events/";
+
+import { init as sentryInit, Integrations } from "@sentry/node";
+import { extraErrorDataIntegration, rewriteFramesIntegration } from "@sentry/integrations";
+import { Prisma } from "@prisma/client";
+import path from "path";
+import { existsSync, readFileSync } from "fs";
 
 
 /**
@@ -65,6 +70,10 @@ export class DiscordClient extends Client {
       .registerEvents();
     new RedisEvents(this)
       .registerEvents();
+
+    // Start Sentry
+    if(process.env["SENTRY_DSN"])
+      this.initSentry();
   }
 
   public async start(token: string) {
@@ -95,6 +104,62 @@ export class DiscordClient extends Client {
         name: `with ${this.guilds.cache.find(g=>g.id===this.config.guildID)?.memberCount} cuties :3`,
         type: ActivityType.Competing,
       }]
+    });
+  }
+
+  private initSentry() {
+    sentryInit({
+      dsn: process.env["SENTRY_DSN"],
+      maxValueLength: 500,
+      integrations: [
+        extraErrorDataIntegration({
+          depth: 5
+        }),
+        rewriteFramesIntegration({
+          iteratee: (frame) => {
+            const absPath = frame.filename;
+            if(!absPath) return frame;
+            // Set the base path as the dist output to match the naming artifact on sentry
+            frame.filename = `/${path.relative(__dirname, absPath).replace(/\\/g, "/")}`;
+            return frame;
+          }
+        }),
+        new Integrations.Prisma({client: this.prisma})
+      ],
+  
+      _experiments: {
+        metricsAggregator: true,
+      },
+    
+      beforeBreadcrumb: (breadcrumb) => {
+        // List of urls to ignore
+        const ignoreUrl = [
+          "https://api.twitch.tv",
+          "https://discord.com",
+          "https://cdn.discordapp.com"
+        ];
+    
+        // Ignore Http Breadcrumbs from the blacklisted url
+        if(breadcrumb.category === "http" && 
+          ignoreUrl.filter(url=>breadcrumb.data?.url.startsWith(url)).length > 0) return null;
+        return breadcrumb;
+      },
+    
+      ignoreErrors: [
+        "ETIMEDOUT",
+        "EADDRINUSE",
+        "ENOTFOUND"
+      ],
+      beforeSend : (evnt, hint) => {
+        if(evnt.tags && evnt.tags["isEval"]) return null;
+    
+        const ex = hint.originalException;
+        if(ex instanceof DiscordAPIError && ex.code === APIErrors.UNKNOWN_INTERACTION) return null;
+        // Somehow prisma bugged and threw this error :/
+        if(ex instanceof Prisma.PrismaClientKnownRequestError && ex.code === "P1017") return null;
+        return evnt;
+      },
+      release: existsSync("commitHash") ? readFileSync("commitHash").toString() : undefined // Pull Release Data
     });
   }
 }
