+++ src/events/index.ts
@@ -1,0 +1,9 @@
+import { AMQPEvents } from "./amqpEvents";
+import { DiscordEvents } from "./discordEvents";
+import { RedisEvents } from "./redisEvents";
+
+export {
+  AMQPEvents,
+  DiscordEvents,
+  RedisEvents
+};
